Improve performance on large projects
We have a moderately-large Go project (~750K lines, split across ~3000 packages) where we want to run gosec as part of CI, but it was taking upwards of 10 minutes for each run, which is above our budget for our whole CI process.

The issue is that gosec is calling packages.Load() N times, once for each package. This means it's regenerating information for common packages many times.

Call packages.Load() once instead, so it can deduplicate this work. This also simplifies the logic for selecting test files, because packages.Load can use the Test flag passed in its Config to determine this.

Calling gosec on github.com/square/p2 with HEAD was taking, on average, 3 minutes to run. With these changes it takes under 3 seconds with a populated build cache, or ~70 seconds with an empty one.

Fixes https://github.com/securego/gosec/issues/406.